Move idempotence logic into inbound adapter module and rename
